requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
idna==3.4
lantz-core==0.5.3
lantz-drivers==0.5.2
lantz-ino==0.5.2
lantz-qt==0.5.3
lantz-sims==0.5.2
lantzdev==0.5.2
pimpmyclass==0.4.3
Pint==0.22
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQt5-stubs==5.15.6.0
pyqt5-tools==5.15.9.3.3
PySignal==1.1.1
python-dotenv==1.0.0
PyVISA==1.13.0
PyYAML==6.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
Serialize==0.2.1
stringparser==0.6
typing_extensions==4.7.1
urllib3==2.0.4
visa==1.0.0
